cleanup code, remove deprecated directives
