+++ src/app/components/trashcan/trashcan.component.html
@@ -12,29 +12,9 @@
                 <mat-icon *ngIf="displayMode === 'gallery'" matTooltip="{{ 'APP.DOCUMENT_LIST.TOOLBAR.LISTVIEW' | translate }}">list</mat-icon>
             </button>
 
-            <!-- <ng-container *ifExperimental="'extensions'"> -->
             <ng-container *ngFor="let entry of actions">
                 <aca-toolbar-action [entry]="entry"></aca-toolbar-action>
             </ng-container>
-            <!-- </ng-container> -->
-
-            <!-- <ng-container *ngIf="!selection.isEmpty">
-                <button
-                    color="primary"
-                    mat-icon-button
-                    [acaPermanentDelete]="selection.nodes"
-                    title="{{ 'APP.ACTIONS.DELETE_PERMANENT' | translate }}">
-                    <mat-icon>delete_forever</mat-icon>
-                </button>
-
-                <button
-                    color="primary"
-                    mat-icon-button
-                    [acaRestoreNode]="selection.nodes"
-                    title="{{ 'APP.ACTIONS.RESTORE' | translate }}">
-                    <mat-icon>restore</mat-icon>
-                </button>
-            </ng-container> -->
         </adf-toolbar>
     </div>
